fix: Issue Fix for Non-owner of the image being able to delete the image

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -151,10 +151,24 @@ public class ImageController {
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
+    //Issue Fix for Non-owner of the image able to delete the image, deleteImageSubmit method is modified to check the logged in User against the owner of the Image
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        String error = "Only the owner of the image can edit the image";
+
+        Image image = imageService.getImage(imageId);
+        String tags = convertTagsToString(image.getTags());
+        model.addAttribute("image", image);
+        model.addAttribute("tags", tags);
+
+        User user = (User) session.getAttribute("loggeduser");
+        if (image.getUser().getId() == user.getId()) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
     }
